# config/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ChessScoreUpdateRequest": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CricketScoreUpdateRequest": {
				"properties": {
					"teamA_runs": {
						"type": "number",
						"format": "double"
					},
					"teamA_overs": {
						"type": "number",
						"format": "double"
					},
					"teamA_wickets": {
						"type": "number",
						"format": "double"
					},
					"teamB_runs": {
						"type": "number",
						"format": "double"
					},
					"teamB_overs": {
						"type": "number",
						"format": "double"
					},
					"teamB_wickets": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_runs",
					"teamA_overs",
					"teamA_wickets",
					"teamB_runs",
					"teamB_overs",
					"teamB_wickets"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FootballScore": {
				"properties": {
					"teamA_score": {
						"type": "number",
						"format": "double"
					},
					"teamB_score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_score",
					"teamB_score"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChessScore": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CricketScore": {
				"properties": {
					"teamA_runs": {
						"type": "number",
						"format": "double"
					},
					"teamA_overs": {
						"type": "number",
						"format": "double"
					},
					"teamA_wickets": {
						"type": "number",
						"format": "double"
					},
					"teamB_runs": {
						"type": "number",
						"format": "double"
					},
					"teamB_overs": {
						"type": "number",
						"format": "double"
					},
					"teamB_wickets": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_runs",
					"teamA_overs",
					"teamA_wickets",
					"teamB_runs",
					"teamB_overs",
					"teamB_wickets"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SquashMenScore": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SquashWomenScore": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TennisMenScore": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TennisWomenScore": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AthleticsEventScore": {
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
...
			"AllScores": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/FootballScore"
					},
					{
						"$ref": "#/components/schemas/ChessScore"
					},
					{
						"$ref": "#/components/schemas/CricketScore"
					},
					{
						"$ref": "#/components/schemas/SquashMenScore"
					},
					{
						"$ref": "#/components/schemas/SquashWomenScore"
					},
					{
						"$ref": "#/components/schemas/TennisMenScore"
					},
					{
						"$ref": "#/components/schemas/TennisWomenScore"
					},
					{
						"$ref": "#/components/schemas/AthleticsEventScore"
					}
				]
			},
			"FootballScoreUpdateRequest": {
				"properties": {
					"teamA_score": {
						"type": "number",
						"format": "double"
					},
					"teamB_score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_score",
					"teamB_score"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"SquashMenScoreUpdateRequest": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SquashWomenScoreUpdateRequest": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Team": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"medals": {
						"properties": {
							"bronze": {
								"type": "number",
								"format": "double"
							},
							"silver": {
								"type": "number",
								"format": "double"
							},
							"gold": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"bronze",
							"silver",
							"gold"
						],
						"type": "object"
					},
					"points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"medals",
					"points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTeamRequest": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TennisMenScoreUpdateRequest": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TennisWomenScoreUpdateRequest": {
				"properties": {
					"teamA_points": {
						"type": "number",
						"format": "double"
					},
					"teamB_points": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"teamA_points",
					"teamB_points"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "inter-iit-scoreboard-server",
		"version": "1.0.0",
		"description": "This is the server for the scoreboard for the Inter-IIT-sports-meet made by Karan Gandhi",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Karan-Gandhi"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/events/chess/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a chess game.",
				"tags": [
					"Events",
					"Chess"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the chess game.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the chess game.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChessScoreUpdateRequest",
								"description": "- The updated score for the chess game."
							}
						}
					}
				}
			}
		},
		"/events/cricket/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a cricket match.",
				"tags": [
					"Events",
					"Cricket"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the match.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CricketScoreUpdateRequest",
								"description": "- The updated score."
							}
						}
					}
				}
			}
		},
...
		"/events/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the score for a specific event.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the event.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated score for the event.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AllScores",
								"description": "- The updated score for the event."
							}
						}
					}
				}
			}
		},
		"/events/football/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Events",
					"Football"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/FootballScoreUpdateRequest"
							}
						}
					}
				}
			}
		},
...
		"/events/squashmen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Events",
					"SquashMen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SquashMenScoreUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/events/squashwomen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Events",
					"SquashWomen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SquashWomenScoreUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/admin/teams": {
			"get": {
				"operationId": "GetAllTeams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Team"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all teams.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddTeam",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Adds a team.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamRequest"
							}
						}
					}
				}
			}
		},
		"/admin/teams/{name}": {
			"get": {
				"operationId": "GetTeamByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/Team"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/teams/{id}": {
			"delete": {
				"operationId": "DeleteTeam",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes a team.",
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the team to be deleted.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/teams/{name}/{medal}": {
			"patch": {
				"operationId": "AddMedal",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "medal",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "name",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/events/tennismen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Events",
					"TennisMen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TennisMenScoreUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/events/tenniswomen/{id}": {
			"put": {
				"operationId": "UpdateScore",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Events",
					"TennisWomen"
				],
				"deprecated": true,
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TennisWomenScoreUpdateRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}